fixed serious problem where forgot to check that only Delete and Backspace keys can delete structures in table. Also changed it so that pressing Delete button will not togle edit mode
git-svn-id: http://hardin/svn/sis/projects/near/near_vis/trunk@44569 ca6fedb5-3f05-0410-a19e-b2f146938ed9

+++ src/edu/jhuapl/near/gui/AbstractStructureMappingControlPanel.java
@@ -305,21 +305,20 @@ public abstract class AbstractStructureMappingControlPanel extends JPanel implem
         {
             public void keyPressed(KeyEvent e)
             {
-                deleteStructure();
+                int keyCode = e.getKeyCode();
+                if (keyCode == KeyEvent.VK_DELETE || keyCode == KeyEvent.VK_BACK_SPACE)
+                    deleteStructure();
             }
         });
     }
 
     private void deleteStructure()
     {
-        editButton.setSelected(false);
-
         int numStructures = structuresTable.getRowCount();
         int idx = structuresTable.getSelectedRow();
         if (idx >= 0 && idx < numStructures)
         {
             structureModel.removeStructure(idx);
-            pickManager.setPickMode(PickManager.PickMode.DEFAULT);
             structureModel.selectStructure(-1);
         }
     }
